Ajusta estructura del script de tokens
