+++ package.json
@@ -29,7 +29,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx src/",
-    "generate-theme": "ts-node --project ./scripts/tsconfig.json ./scripts/generateDesignTokens && npx eslint src/styles/designTokens --fix"
+    "generate-theme": "ts-node --project ./scripts/tsconfig.json ./scripts/generateDesignTokens/generateDesignTokens && npx eslint src/styles/designTokens --fix"
   },
   "browserslist": {
     "production": [
