--- src/styles/designTokens/uiTokens.ts
@@ -1,69 +1,0 @@
-/* eslint-disable max-len */
-export const UiTokens = {
-    "borderRadius": {
-        "size0": "0px",
-        "size4": "4px",
-        "size8": "8px",
-        "size16": "16px",
-        "size20": "20px",
-        "size32": "32px",
-        "size48": "48px",
-        "size60": "60px",
-        "size76": "76px",
-        "sizeTotal": "100%",
-    },
-    "borderWidth": {
-        "size1": "1px",
-        "size2": "2px",
-        "size3": "3px",
-        "size4": "4px",
-        "size5": "5px",
-        "size6": "6px",
-        "size7": "7px",
-        "size8": "8px",
-        "size9": "9px",
-        "size10": "10px",
-    },
-    "spacing": {
-        "size0": "0",
-        "size1": "1px",
-        "size2": "2px",
-        "size4": "4px",
-        "size8": "8px",
-        "size12": "12px",
-        "size16": "16px",
-        "size20": "20px",
-        "size24": "24px",
-        "size28": "28px",
-        "size32": "32px",
-        "size36": "36px",
-        "size40": "40px",
-        "size44": "44px",
-        "size48": "48px",
-        "size52": "52px",
-        "size56": "56px",
-        "size60": "60px",
-        "size68": "68px",
-        "size72": "72px",
-        "size80": "80px",
-        "size88": "88px",
-        "size96": "96px",
-        "size112": "112px",
-        "size128": "128px",
-        "size160": "160px",
-        "size208": "208px",
-        "size256": "265px",
-    },
-    "shadows": {
-        "shadow4": "0px 25px 7px 0px rgba(0,0,0,0),0px 16px 6px 0px rgba(0,0,0,0.01),0px 9px 5px 0px rgba(0,0,0,0.03),0px 4px 4px 0px rgba(0,0,0,0.05),0px 1px 2px 0px rgba(0,0,0,0.06),0px 0px 0px 0px rgba(0,0,0,0.06)",
-        "shadow6": "0px 33px 9px 0px rgba(0,0,0,0),0px 21px 9px 0px rgba(0,0,0,0.01),0px 12px 7px 0px rgba(0,0,0,0.03),0px 5px 5px 0px rgba(0,0,0,0.05),0px 1px 3px 0px rgba(0,0,0,0.06),0px 0px 0px 0px rgba(0,0,0,0.06)",
-        "shadow8": "0px 43px 12px 0px rgba(0,0,0,0),0px 27px 11px 0px rgba(0,0,0,0.01),0px 15px 9px 0px rgba(0,0,0,0.03),0px 7px 7px 0px rgba(0,0,0,0.05),0px 2px 4px 0px rgba(0,0,0,0.06),0px 2px 4px 0px rgba(0,0,0,0.06)",
-        "shadow12": "0px 67px 19px 0px rgba(0,0,0,0),0px 43px 17px 0px rgba(0,0,0,0.01),0px 24px 14px 0px rgba(0,0,0,0.03),0px 11px 11px 0px rgba(0,0,0,0.05),0px 3px 6px 0px rgba(0,0,0,0.06),0px 0px 0px 0px rgba(0,0,0,0.06)",
-        "shadow16": "0px 91px 26px 0px rgba(0,0,0,0),0px 58px 23px 0px rgba(0,0,0,0.01),0px 33px 20px 0px rgba(0,0,0,0.03),0px 15px 15px 0px rgba(0,0,0,0.05),0px 4px 8px 0px rgba(0,0,0,0.06),0px 0px 0px 0px rgba(0,0,0,0.06)",
-        "shadow24": "0px 133px 37px 0px rgba(0,0,0,0),0px 85px 34px 0px rgba(0,0,0,0.01),0px 48px 29px 0px rgba(0,0,0,0),0px 21px 21px 0px rgba(0,0,0,0.05),0px 5px 12px 0px rgba(0,0,0,0.06),0px 0px 0px 0px rgba(0,0,0,0.06)",
-    },
-    "innerShadows": {
-        "shadow4": "inner 0px 25px 7px 0px rgba(0,0,0,0),inner 0px 16px 6px 0px rgba(0,0,0,0.01),inner 0px 9px 5px 0px rgba(0,0,0,0.03),inner 0px 4px 4px 0px rgba(0,0,0,0.05),inner 0px 1px 2px 0px rgba(0,0,0,0.06),inner 0px 0px 0px 0px rgba(0,0,0,0.06)",
-        "shadow12": "inner 0px 66px 19px 0px rgba(0,0,0,0),inner 0px 42px 17px 0px rgba(0,0,0,0.01),inner 0px 24px 14px 0px rgba(0,0,0,0.03),inner 0px 11px 11px 0px rgba(0,0,0,0.05),inner 0px 3px 6px 0px rgba(0,0,0,0.06),inner 0px 0px 0px 0px rgba(0,0,0,0.06)",
-    },
-} as const
